+++ homeworks/03-data-structure/README.md
@@ -31,5 +31,5 @@ All files             |     100 |      100 |     100 |     100 |
  merge-sort           |     100 |      100 |     100 |     100 |                   
   mergeSort.ts        |     100 |      100 |     100 |     100 |                   
  utils                |     100 |      100 |     100 |     100 |                   
-  commonUtils.ts      |     100 |      100 |     100 |     100 |                   
+  common.ts           |     100 |      100 |     100 |     100 |                   
 ----------------------|---------|----------|---------|---------|-------------------
