+++ homeworks/03-data-structure/src/utils/common.ts
